requirements.txt
torch
torchvision
torchaudio
facenet-pytorch
opencv-python
numpy >= 1.23
tqdm
pyserial
